+++ src/geos-common.h
@@ -66,7 +66,7 @@ SEXP geos_common_tree_xptr(GEOSSTRtree* geometry, SEXP geom, SEXP indices);
 // check geometry coming out of an externalptr
 #define GEOS_CHECK_GEOMETRY(geometry, i)  \
   if (geometry == NULL) {   \
-    Rf_error("External pointer is not valid [i=%d]", i + 1);  \
+    Rf_error("External pointer is not valid [i=%ld]", (long)i + 1);  \
   }
